+++ api/src/main/java/net/jqwik/api/Falsifier.java
@@ -16,46 +16,12 @@ public interface Falsifier<T> {
 	TryExecutionResult execute(T t);
 
 	@API(status = INTERNAL)
-	default FalsificationResult<T> falsify(Shrinkable<T> candidate) {
-		try {
-			TryExecutionResult executionResult = execute(candidate.value());
-			switch (executionResult.status()) {
-				case FALSIFIED:
-					return FalsificationResult.falsified(candidate, executionResult.throwable().orElse(null));
-				case INVALID:
-					return FalsificationResult.filtered(candidate);
-				default:
-					return FalsificationResult.notFalsified(candidate);
-			}
-		} catch (Throwable throwable) {
-			return FalsificationResult.falsified(candidate, throwable);
-		}
-	}
-
-	@API(status = INTERNAL)
 	default Falsifier<T> withFilter(Predicate<T> filter) {
 		return value -> {
 			if (!filter.test(value)) {
 				return TryExecutionResult.invalid();
 			}
 			return execute(value);
-		};
-	}
-
-	@API(status = INTERNAL)
-	default <U> Falsifier<U> map(Function<U, T> mapper) {
-		return value -> {
-			T other = null;
-			try {
-				other = mapper.apply(value);
-			} catch (Throwable throwable) {
-				// Ignore exceptions during shrinking
-				if (throwable instanceof OutOfMemoryError) {
-					throw throwable;
-				}
-				return TryExecutionResult.invalid();
-			}
-			return Falsifier.this.execute(other);
 		};
 	}